move optional space breakup of long table cells to Presentation XML: https://github.com/metanorma/metanorma-ogc/issues/559
