+++ lib/isodoc/presentation_function/block.rb
@@ -1,12 +1,11 @@
-require_relative "./image"
-require_relative "./sourcecode"
+require_relative "image"
+require_relative "sourcecode"
 require "rouge"
 
 module IsoDoc
   class PresentationXMLConvert < ::IsoDoc::Convert
     def lower2cap(text)
-      return text if /^[[:upper:]][[:upper:]]/.match?(text)
-
+      /^[[:upper:]][[:upper:]]/.match?(text) and return text
       text.capitalize
     end
 
@@ -89,31 +88,8 @@ module IsoDoc
       prefix_name(elem, block_delim, l10n("#{@i18n.box} #{n}"), "name")
     end
 
-    def recommendation(docxml)
-      docxml.xpath(ns("//recommendation")).each do |f|
-        recommendation1(f, lower2cap(@i18n.recommendation))
-      end
-    end
-
-    def requirement(docxml)
-      docxml.xpath(ns("//requirement")).each do |f|
-        recommendation1(f, lower2cap(@i18n.requirement))
-      end
-    end
-
-    def permission(docxml)
-      docxml.xpath(ns("//permission")).each do |f|
-        recommendation1(f, lower2cap(@i18n.permission))
-      end
-    end
-
-    def recommendation1(elem, type)
-      lbl = @reqt_models.model(elem["model"])
-        .recommendation_label(elem, type, xrefs)
-      prefix_name(elem, "", l10n(lbl), "name")
-    end
-
     def table(docxml)
+      table_long_strings_cleanup(docxml)
       docxml.xpath(ns("//table")).each { |f| table1(f) }
     end
 
@@ -123,6 +99,74 @@ module IsoDoc
       n = @xrefs.anchor(elem["id"], :label, false)
       prefix_name(elem, block_delim, l10n("#{lower2cap @i18n.table} #{n}"),
                   "name")
+    end
+
+    def table_long_strings_cleanup(docxml)
+      @break_up_urls_in_tables or return
+      docxml.xpath("//td | //th").each do |d|
+        d.traverse do |n|
+          n.text? or next
+          ret = break_up_long_str(n.text)
+          n.content = ret
+        end
+      end
+    end
+
+    LONGSTR_THRESHOLD = 10
+    LONGSTR_NOPUNCT = 2
+
+    def break_up_long_str(text)
+      /^\s*$/.match?(text) and return text
+      text.split(/(?=\s)/).map do |w|
+        if /^\s*$/.match(text) || (w.size < LONGSTR_THRESHOLD) then w
+        else
+          w.scan(/.{,#{LONGSTR_THRESHOLD}}/o).map.with_index do |w1, i|
+            w1.size < LONGSTR_THRESHOLD ? w1 : break_up_long_str1(w1, i + 1)
+          end.join
+        end
+      end.join
+    end
+
+    STR_BREAKUP_RE = %r{
+     (?<=[=_—–\u2009→?+;]) | # break after any of these
+     (?<=[,.:])(?!\d) | # break on punct only if not preceding digit
+     (?<=[>])(?![>]) | # > not >->
+     (?<=[\]])(?![\]]) | # ] not ]-]
+     (?<=//) | # //
+     (?<=[/])(?![/]) | # / not /-/
+     (?<![<])(?=[<]) | # < not <-<
+     (?<=\p{L})(?=[(\{\[]\p{L}) # letter and bracket, followed by letter
+    }x.freeze
+
+
+    CAMEL_CASE_RE = %r{
+      (?<=\p{Ll}\p{Ll})(?=\p{Lu}\p{Ll}\p{Ll}) # 2 lowerc / upperc, 2 lowerc
+    }x.freeze
+
+    # break on punct every LONGSTRING_THRESHOLD chars, with zero width space
+    # if punct fails, try break on camel case, with soft hyphen
+    # break regardless every LONGSTRING_THRESHOLD * LONGSTR_NOPUNCT,
+    # with soft hyphen
+    def break_up_long_str1(text, iteration)
+      s, separator = break_up_long_str2(text)
+      if s.size == 1 # could not break up
+        (iteration % LONGSTR_NOPUNCT).zero? and
+          text += "\u00ad" # force soft hyphen
+        text
+      else
+        s[-1] = "#{separator}#{s[-1]}"
+        s.join
+      end
+    end
+
+    def break_up_long_str2(text)
+      s = text.split(STR_BREAKUP_RE, -1)
+      separator = "\u200b"
+      if s.size == 1
+        s = text.split(CAMEL_CASE_RE)
+        separator = "\u00ad"
+      end
+      [s, separator]
     end
 
     # we use this to eliminate the semantic amend blocks from rendering
@@ -165,26 +209,6 @@ module IsoDoc
 
     def ol_label(elem)
       elem["label"] = @xrefs.anchor(elem["id"], :label, false)
-    end
-
-    def requirement_render_preprocessing(docxml); end
-
-    REQS = %w(requirement recommendation permission).freeze
-
-    def requirement_render(docxml)
-      requirement_render_preprocessing(docxml)
-      REQS.each do |x|
-        REQS.each do |y|
-          docxml.xpath(ns("//#{x}//#{y}")).each { |r| requirement_render1(r) }
-        end
-      end
-      docxml.xpath(ns("//requirement | //recommendation | //permission"))
-        .each { |r| requirement_render1(r) }
-    end
-
-    def requirement_render1(node)
-      node.replace(@reqt_models.model(node["model"])
-        .requirement_render1(node))
     end
 
     def source(docxml)
