+++ spec/isodoc/cleanup_spec.rb
@@ -592,258 +592,4 @@ RSpec.describe IsoDoc do
     expect(xmlpp(IsoDoc::HtmlConvert.new({})
       .cleanup(Nokogiri::XML(input)).to_s)).to be_equivalent_to xmlpp((html))
   end
-
-  it "does not break up very long strings in tables by default" do
-    input = <<~INPUT
-        <html xmlns:epub="http://www.idpf.org/2007/ops">
-      <head>
-        <title>test</title>
-        <body lang="EN-US" link="blue" vlink="#954F72">
-          <div class="WordSection1">
-            <p>&#160;</p>
-          </div>
-          <br clear="all" class="section"/>
-          <div class="WordSection2">
-            <br clear="all" style="mso-special-character:line-break;page-break-before:always"/>
-            <div>
-              <h1 class="ForewordTitle">Foreword</h1>
-              <p class="TableTitle" align="center">
-                <b>Table 1&#160;&#8212; Repeatability and reproducibility of husked rice yield</b>
-              </p>
-              <table id="tableD-1" class="MsoISOTable" border="1" cellspacing="0" cellpadding="0">
-                <thead>
-                  <tr>
-                    <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">Description</td>
-                    <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">Description</td>
-                    <td align="center" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;">Rice sample</td>
-                  </tr>
-                  </thead>
-                  <tbody>
-                  <tr>
-                    <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">http://www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/BBBBBBBBBBBBBBBBBBBBBBBBBBBB</td>
-                    <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">http://www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABBBBBBBBBBBBBBBBBBBBBBBBBBBB</td>
-                    <td align="center" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;">www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABBBBBBBBBBBBBBBBBBBBBBBBBBBB</td>
-                  </tr>
-                  </tbody>
-                  </table>
-                  </div>
-                  </div>
-                  </body>
-                  </html>
-    INPUT
-    output = <<~OUTPUT
-                           <?xml version='1.0'?>
-      <html xmlns:epub='http://www.idpf.org/2007/ops'>
-        <head>
-          <title>test</title>
-          <body lang='EN-US' link='blue' vlink='#954F72'>
-            <div class='WordSection1'>
-              <p>&#xA0;</p>
-            </div>
-            <br clear='all' class='section'/>
-            <div class='WordSection2'>
-              <br clear='all' style='mso-special-character:line-break;page-break-before:always'/>
-              <div>
-                <h1 class='ForewordTitle'>Foreword</h1>
-                <p class='TableTitle' align='center'>
-                  <b>Table 1&#xA0;&#x2014; Repeatability and reproducibility of husked rice yield</b>
-                </p>
-                <table id='tableD-1' class='MsoISOTable' border='1' cellspacing='0' cellpadding='0'>
-                  <thead>
-                    <tr>
-                      <td align='left' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;'>Description</td>
-                      <td align='left' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;'>Description</td>
-                      <td align='center' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;'>Rice sample</td>
-                    </tr>
-                    </thead>
-                    <tbody>
-                      <tr>
-                        <td align='left' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;'>
-                        http://www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/BBBBBBBBBBBBBBBBBBBBBBBBBBBB
-                        </td>
-                        <td align='left' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;'>
-                        http://www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABBBBBBBBBBBBBBBBBBBBBBBBBBBB </td>
-                        <td align='center' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;'>
-                        www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABBBBBBBBBBBBBBBBBBBBBBBBBBBB
-                        </td>
-                      </tr>
-                    </tbody>
-                </table>
-              </div>
-            </div>
-          </body>
-        </head>
-      </html>
-    OUTPUT
-    expect(xmlpp(IsoDoc::HtmlConvert.new({})
-      .cleanup(Nokogiri::XML(input)).to_s)).to be_equivalent_to xmlpp((output))
-    output = <<~OUTPUT
-                           <?xml version='1.0'?>
-      <html xmlns:epub='http://www.idpf.org/2007/ops'>
-        <head>
-          <title>test</title>
-          <body lang='EN-US' link='blue' vlink='#954F72'>
-            <div class='WordSection1'>
-              <p>&#xA0;</p>
-            </div>
-            <br clear='all' class='section'/>
-            <div class='WordSection2'>
-              <br clear='all' style='mso-special-character:line-break;page-break-before:always'/>
-              <div>
-                <h1 class='ForewordTitle'>Foreword</h1>
-                <p class='TableTitle' align='center'>
-                  <b>Table 1&#xA0;&#x2014; Repeatability and reproducibility of husked rice yield</b>
-                </p>
-                <table id='tableD-1' class='MsoISOTable' border='1' cellspacing='0' cellpadding='0'>
-                  <thead>
-                    <tr>
-                      <td align='left' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;'>Description</td>
-                      <td align='left' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;'>Description</td>
-                      <td align='center' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;'>Rice sample</td>
-                    </tr>
-                    </thead>
-                    <tbody>
-                      <tr>
-                        <td align='left' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;'>
-                          http://www.example.com/&amp;AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/BBBBBBBBBBBBBBBBBBBBBBBBBBBB
-                        </td>
-                        <td align='left' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;'>
-                          http://www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABBBBBBBBBBBBBBBBBBBBBBBBBBBB
-                        </td>
-                        <td align='center' style='border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;'>
-                          www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABBBBBBBBBBBBBBBBBBBBBBBBBBBB
-                        </td>
-                      </tr>
-                    </tbody>
-                </table>
-              </div>
-            </div>
-          </body>
-        </head>
-      </html>
-    OUTPUT
-    expect(xmlpp(IsoDoc::WordConvert.new({})
-      .cleanup(Nokogiri::XML(input)).to_s)).to be_equivalent_to xmlpp((output))
-  end
-
-  it "breaks up very long strings in tables on request" do
-    input = <<~INPUT
-        <html xmlns:epub="http://www.idpf.org/2007/ops">
-      <head>
-        <title>test</title>
-        <body lang="EN-US" link="blue" vlink="#954F72">
-          <div class="WordSection1">
-            <p>&#160;</p>
-          </div>
-          <br clear="all" class="section"/>
-          <div class="WordSection2">
-            <br clear="all" style="mso-special-character:line-break;page-break-before:always"/>
-            <div>
-              <h1 class="ForewordTitle">Foreword</h1>
-              <p class="TableTitle" align="center">
-                <b>Table 1&#160;&#8212; Repeatability and reproducibility of husked rice yield</b>
-              </p>
-              <table id="tableD-1" class="MsoISOTable" border="1" cellspacing="0" cellpadding="0">
-                <thead>
-                  <tr>
-                    <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">Description</td>
-                    <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">Description</td>
-                    <td align="center" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;">Rice sample</td>
-                  </tr>
-                  </thead>
-                  <tbody>
-                  <tr>
-                    <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">http://www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/BBBBBBBBBBBBBBBBBBBBBBBBBBBB</td>
-                    <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">http://www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABBBBBBBBBBBBBBBBBBBBBBBBBBBB</td>
-                    <td align="center" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;">www.example.com/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABBBBBBBBBBBBBBBBBBBBBBBBBBBB</td>
-                  </tr>
-                  </tbody>
-                  </table>
-                  </div>
-                  </div>
-                  </body>
-                  </html>
-    INPUT
-    html = <<~OUTPUT
-           <html xmlns:epub="http://www.idpf.org/2007/ops">
-       <head>
-         <title>test</title>
-         <body lang="EN-US" link="blue" vlink="#954F72">
-           <div class="WordSection1">
-             <p>&#xA0;</p>
-           </div>
-           <br clear="all" class="section"/>
-           <div class="WordSection2">
-             <br clear="all" style="mso-special-character:line-break;page-break-before:always"/>
-             <div>
-               <h1 class="ForewordTitle">Foreword</h1>
-               <p class="TableTitle" align="center">
-                 <b>Table 1&#xA0;&#x2014; Repeatability and reproducibility of husked rice yield</b>
-               </p>
-               <table id="tableD-1" class="MsoISOTable" border="1" cellspacing="0" cellpadding="0">
-                 <thead>
-                   <tr>
-                     <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">Description</td>
-                     <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">Description</td>
-                     <td align="center" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;">Rice sample</td>
-                   </tr>
-                   </thead>
-                   <tbody>
-                   <tr>
-                                        <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">http://&#x200B;www.example.&#x200B;com/&#x200B;AAAAAAAAAAAAAAAAA&#x200B;AAAAAAAAAAAAAAAAAAAA&#x200B;AAAAAAAA/&#x200B;BBBBBBBBBBB&#x200B;BBBBBBBBBBBBBBBBB</td>
-                     <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">http://&#x200B;www.example.&#x200B;com/&#x200B;AAAAAAAAAAAAAAAAA&#x200B;AAAAAAAAAAAAAAAAAAAA&#x200B;AAAAAAAABBBBBBBBBBBB&#x200B;BBBBBBBBBBBBBBBB</td>
-                     <td align="center" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;">www.&#x200B;example.com/&#x200B;AAAAAAAAAAAAAAAAAAAAAAAA&#x200B;AAAAAAAAAAAAAAAAAAAA&#x200B;ABBBBBBBBBBBBBBBBBBB&#x200B;BBBBBBBBB</td>
-                   </tr>
-                   </tbody>
-                   </table>
-                   </div>
-                   </div>
-                   </body>
-                   </head>
-       </html>
-    OUTPUT
-    doc = <<~OUTPUT
-           <html xmlns:epub="http://www.idpf.org/2007/ops">
-       <head>
-         <title>test</title>
-         <body lang="EN-US" link="blue" vlink="#954F72">
-           <div class="WordSection1">
-             <p>&#xA0;</p>
-           </div>
-           <br clear="all" class="section"/>
-           <div class="WordSection2">
-             <br clear="all" style="mso-special-character:line-break;page-break-before:always"/>
-             <div>
-               <h1 class="ForewordTitle">Foreword</h1>
-               <p class="TableTitle" align="center">
-                 <b>Table 1&#xA0;&#x2014; Repeatability and reproducibility of husked rice yield</b>
-               </p>
-               <table id="tableD-1" class="MsoISOTable" border="1" cellspacing="0" cellpadding="0">
-                 <thead>
-                   <tr>
-                     <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">Description</td>
-                     <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">Description</td>
-                     <td align="center" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;">Rice sample</td>
-                   </tr>
-                   </thead>
-                   <tbody>
-                   <tr>
-                                          <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">http://&#x200B;www.example.&#x200B;com/&#x200B;AAAAAAAAAAAAAAAAA&#x200B;AAAAAAAAAAAAAAAAAAAA&#x200B;AAAAAAAA/&#x200B;BBBBBBBBBBB&#x200B;BBBBBBBBBBBBBBBBB</td>
-                     <td align="left" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.5pt;">http://&#x200B;www.example.&#x200B;com/&#x200B;AAAAAAAAAAAAAAAAA&#x200B;AAAAAAAAAAAAAAAAAAAA&#x200B;AAAAAAAABBBBBBBBBBBB&#x200B;BBBBBBBBBBBBBBBB</td>
-                     <td align="center" style="border-top:solid windowtext 1.5pt;border-bottom:solid windowtext 1.0pt;">www.&#x200B;example.com/&#x200B;AAAAAAAAAAAAAAAAAAAAAAAA&#x200B;AAAAAAAAAAAAAAAAAAAA&#x200B;ABBBBBBBBBBBBBBBBBBB&#x200B;BBBBBBBBB</td>
-                   </tr>
-                   </tbody>
-                   </table>
-                   </div>
-                   </div>
-                   </body>
-                   </head>
-       </html>
-    OUTPUT
-    # no xmlpp, it converts ZWSP to full space
-    expect(IsoDoc::HtmlConvert.new({ breakupurlsintables: "true" })
-      .cleanup(Nokogiri::XML(input)).to_s).to be_equivalent_to (html)
-    expect(IsoDoc::WordConvert.new({ breakupurlsintables: "true" })
-      .cleanup(Nokogiri::XML(input)).to_s).to be_equivalent_to (doc)
-  end
 end
